feat: initial implementation of alfred-notion
Plenty features are also done, such as the ability to:

 - Open current week/day pages
 - Update/Add tags (cached to file)
 - Add Tasks (with tag and status selections)
 - Add Wins (with tag selections)
 - Search tasks (open and change status)

Configurations are handled with a data/config.json file. This is to work
around the limitations of Alfred env vars (I don't want to expose mine
in the info.plist).

There is much to improve, but it is a start.