refactor: remove unused code, and leave reference material comment
Originally, I took the source from alfred-notion for the notion_api.py.

Not everything is being used in this project, so I cleaned it up. I also
left a note from where I got it from, and how it could be cleaned up in
the future (i.e., single package both projects use).