refactor: duplicated code out into new repository/package notion-scripts
To share similar code between this project and alfred-notion (along with
other future projects) a new repository notion-scripts was created.

This commit removes the duplicated code and uses the new package. The
config file isn't needed anymore, as the default config in
notion-scripts is the same as what this project needed.

Some refactoring was done in the new package to be more class-based.

Some additional refactors were done to better accomodate the new
changes.

The transition_tasks scheduled job is using the same behaviour of
creating a new notion client for each iteration.