refactor: duplicated code out into new repository/package notion-scripts
To share similar code between this project and notion-heroku (along with
other future projects) a new repository notion-scripts was created.

This commit removes the duplicated code and uses the new package.

Some refactoring was done in the new package to be more class-based.

The notion_api now defines the notion api usage (with the alfred based
config being used instead of the default from notion-scripts).