fix: remove workflow release approach (data leaked)
So turns out that my approach of not having the env vars in the alfred
workflow itself (as it appears in the info.plist even with the 'do not
export') backfired.

I have configurations in the `data.config.json` which are not committed
to the repository. Althought, when you export the alfred workflow itself
it'll actually dump everything that was in the directory (which included
my actual `data.config.json` file (oops).

Fortunately, my notion_token has cycled and that the rest of the values
are not that important on their own.

It would be possible to keep things as is and workaround it... but it is
a more involved exporting process. I've decided to removed the exported
workflow from GitHub for the time being, and update the README to
address these changes with respect to how to install this workflow.

Thanks @haakonstorm (haakon storm heen) for notifying me about this :)