fix: open in app_url functionality
It turns out you have to keep the `www.notion.so` part in the url for
the page to open in the desktop application.

This changed a few months back and I even confirmed with the Notion
support team that this was an issue. Turns out things just changed. This
change now make it behave as expected.