fix: make 'server' directory Heroku deployable (moving files to root)
Heroku expects the root directory to contain the details on how to
deploy the application. In this case we have a monorepo, but we want the
'server' directory to be deployed.

To further complicate the situation the 'shared' directory on the root
level that is needed for the server due to how we have the dependencies
laid out.

The temporary solution is to move the required files up from the
'server' directory into the root. This way we heroku will deploy the
application (although the rest of the source files are still present).

One approach I want to try later is to have this somehow handled with
Docker, and just have a deployable container for the application.